+++ python/dgl/contrib/__init__.py
@@ -1,5 +1,4 @@
 from . import sampling
 from . import graph_store
 from .dis_kvstore import KVClient, KVServer
-from .dist_graph import DistGraphServer, DistGraph
 from .dis_kvstore import read_ip_config
