+++ python/dgl/distributed/__init__.py
@@ -1,0 +1,2 @@
+from .dist_graph import DistGraphServer, DistGraph
+from .partition import partition_graph, load_partition
